Moving glad into dist folder
